+++ R/tidy_xlsx.R
@@ -14,7 +14,7 @@
 #'
 #' @details
 #' A cell has two 'values': its content, and sometimes also a formula.  It also
-#' has formatting applied at the 'style' level, which can be locally overriden.
+#' has formatting applied at the 'style' level, which can be locally overridden.
 #'
 #' \subsection{content}{
 #'   Depending on the cell, the content may be a numeric value such as 365 or
@@ -31,7 +31,7 @@
 #'   of the formula the last time it was evaluated.
 #'
 #'   Certain groups of cells may share a formula that differs only by addresses
-#'   refered to in the formula; such groups are identified by an index, the
+#'   referred to in the formula; such groups are identified by an index, the
 #'   'formula_group'.  The xlsx (Excel) file format only records the formula
 #'   against one cell in any group.  It is planned for \code{tidy_xlsx} to parse
 #'   such formulas and copy them to the other cells in a group, making the
@@ -74,7 +74,7 @@
 #' \describe{
 #'   \item{address}{The cell address in A1 notation.}
 #'   \item{row}{The row number of a cell address (integer).}
-#'   \item{col}{The column numer of a cell address (integer).}
+#'   \item{col}{The column number of a cell address (integer).}
 #'   \item{content}{The content of a cell before type inference (see
 #'   'Details').}
 #'   \item{formula}{The formula in a cell (see 'Details').}
@@ -160,14 +160,14 @@ tidy_xlsx <- function(path, sheets = NA) {
   if (anyNA(sheets)) {
     if (length(sheets) > 1) {
       warning("Argument 'sheets' included NAs, which were discarded.")
-      sheets = sheets[!is.na(sheets)]
+      sheets <- sheets[!is.na(sheets)]
       if (length(sheets) == 0) {
         stop("All elements of argument 'sheets' were discarded.")
       }
     } else {
-      sheets = all_sheets$index
+      sheets <- all_sheets$order
     }
   }
   sheets <- standardise_sheet(sheets, all_sheets)
-  xlsx_read_(path, sheets$index, sheets$name, sheets$comments_path)
+  xlsx_read_(path, sheets$sheet_path, sheets$name, sheets$comments_path)
 }