+++ src/xlsxsheet.h
@@ -11,9 +11,11 @@ class xlsxsheet {
 
     std::string name_;
 
+    unsigned long long int i_; // cell counter
     double defaultRowHeight_;
     double defaultColWidth_;
     std::vector<double> colWidths_;
+    std::vector<double> rowHeights_;
     xlsxbook& book_; // reference to parent workbook
     Rcpp::List information_; // Wrapper for variables returned to R
 
@@ -46,7 +48,11 @@ class xlsxsheet {
     Rcpp::CharacterVector style_format_;    // cellXfs xfId links to cellStyleXfs entry
     Rcpp::IntegerVector   local_format_id_; // cell 'c' links to cellXfs entry
 
-    xlsxsheet(const int& sheetindex, xlsxbook& book, Rcpp::String comments_path);
+    xlsxsheet(
+        const std::string& name,
+        const std::string& sheet_path,
+        xlsxbook& book,
+        Rcpp::String comments_path);
     Rcpp::List& information();       // Cells contents and styles DF wrapped in list
 
     void cacheDefaultRowColDims(rapidxml::xml_node<>* worksheet);
@@ -55,6 +61,7 @@ class xlsxsheet {
     void cacheComments(Rcpp::String comments_path);
     void initializeColumns(rapidxml::xml_node<>* sheetData);
     void parseSheetData(rapidxml::xml_node<>* sheetData);
+    void appendComments();
 
 };